back/pyproject.toml
[tool.poetry]
name = "backpoetry"
version = "0.1.0"
description = ""
authors = ["ctomrp <cristian.rico.palma@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
django = "^4.2.5"
djangorestframework = "^3.14.0"
django-cors-headers = "^4.2.0"
coreapi = "^2.3.3"
pyjwt = "^2.8.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
